+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 - Fix #93: Fixed sample data for galera compatibility
 - Enh #95: Allow update content topics
 - Enh #96: Change icon in sidebar menu
+- Fix #10: Fix yii console Help without installed Rest module
 
 
 1.0.1 (February 5, 2021)